pyproject.toml
[tool.poetry]
name = "todolist"
version = "0.1.0"
description = "graduation project"
authors = ["Yuriy S. <trupakenator@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.3"
django-environ = "^0.9.0"
psycopg2-binary = "^2.9.5"
djangorestframework = "^3.14.0"
django-cors-headers = "^3.13.0"
social-auth-app-django = "^5.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
